Move ConnectionHelper and DatabaseHelper into dedicated support files
